+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/Timestampable.java
@@ -20,4 +20,6 @@ public interface Timestampable<T> {
 
     T usingTimestamps();
 
+    T withoutUsingTimestamps();
+
 }