+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/PodOperationsImpl.java
@@ -664,5 +664,10 @@ public class PodOperationsImpl extends HasMetadataOperation<Pod, PodList, Doneab
     return new PodOperationsImpl(getContext().withTimestamps(true));
   }
 
+  @Override
+  public BytesLimitTerminateTimeTailPrettyLoggable<String, LogWatch> withoutUsingTimestamps() {
+    return new PodOperationsImpl(getContext().withTimestamps(false));
+  }
+
 }